+++ modules/selenoid/src/test/java/integration/FileDownloadTest.java
@@ -1,55 +1,45 @@
 package integration;
 
 import com.codeborne.selenide.Configuration;
-import com.codeborne.selenide.FileDownloadMode;
-import org.junit.jupiter.api.AfterEach;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
 
-import java.io.File;
 import java.io.IOException;
-import org.openqa.selenium.net.NetworkUtils;
 
 import static com.codeborne.selenide.FileDownloadMode.FOLDER;
 import static com.codeborne.selenide.FileDownloadMode.HTTPGET;
 import static com.codeborne.selenide.FileDownloadMode.PROXY;
-import static com.codeborne.selenide.Selectors.byText;
-import static com.codeborne.selenide.Selenide.$;
-import static com.codeborne.selenide.Selenide.open;
-import static com.codeborne.selenide.files.FileFilters.withExtension;
-import static java.nio.charset.StandardCharsets.UTF_8;
-import static org.apache.commons.io.FileUtils.readFileToString;
-import static org.assertj.core.api.Assertions.assertThat;
+import static integration.SelenoidSetup.checkDownload;
+import static integration.SelenoidSetup.resetSelenoidSettings;
 
-public class FileDownloadTest extends ITest {
+@ExtendWith(SelenoidSetup.class)
+public class FileDownloadTest {
   @Test
-  void downloadViaHttpGet() throws IOException {
+  void downloadFileInSelenoid_using_httpGet() throws IOException {
     Configuration.proxyEnabled = false;
     Configuration.fileDownload = HTTPGET;
     checkDownload();
   }
 
   @Test
-  void downloadViaFolder() throws IOException {
+  void downloadFileInSelenoid_using_folder() throws IOException {
     Configuration.proxyEnabled = false;
     Configuration.fileDownload = FOLDER;
     checkDownload();
   }
 
   @Test
-  void downloadViaProxy() throws IOException {
+  void downloadFileInSelenoid_using_proxy() throws IOException {
     Configuration.proxyEnabled = true;
-    Configuration.proxyHost = new NetworkUtils().getNonLoopbackAddressOfThisMachine();
     Configuration.fileDownload = PROXY;
     checkDownload();
   }
 
-  private void checkDownload() throws IOException {
-    for (int i = 0; i < 2; i++) {
-      open("/file_download.html");
-      File file = $(byText("some-file.txt")).download(withExtension("txt"));
-      assertThat(file).hasName("some-file.txt");
-      assertThat(readFileToString(file, UTF_8)).startsWith("{\\rtf");
-    }
+  @Test
+  void downloadFileInLocalBrowser() throws IOException {
+    resetSelenoidSettings();
+    Configuration.headless = true;
+    Configuration.fileDownload = FOLDER;
+    checkDownload();
   }
 }